Add SER and relevant metrics for hap
