--- src/utils.py
@@ -1,1846 +1,0 @@
-import argparse
-import ast
-import os
-import shutil
-import subprocess
-import time
-import warnings
-
-import numpy as np
-
-import src.tinypeel.tinypeel as tinypeel
-
-
-ACCURACY_TEST_ROOT = os.path.join("tests", "accuracy_tests")
-SIMULATION_PARAMETERS_FILE = os.path.join(
-    ACCURACY_TEST_ROOT, "simulation_parameters.txt"
-)
-SIMULATION_FIXTURE_DIR = os.path.join(ACCURACY_TEST_ROOT, "sim_for_alphapeel_accu_test")
-
-DEFAULT_ALPHA_PEEL_ARGS = {
-    "n_cycle": "5",
-    "n_thread": "6",
-    "geno_threshold": ".1",
-    "hap_threshold": ".1",
-    "geno": None,
-    "hap": None,
-    "seg_prob": None,
-    "geno_prob": None,
-    "phased_geno_prob": None,
-}
-
-BASE_ACCURACY_FILES = [
-    "dosage",
-    "geno_0.333",
-    "hap_0.5",
-    "geno_prob",
-    "phased_geno_prob",
-]
-METAFOUNDER_ACCURACY_FILES = [
-    "dosage",
-    "geno_prob",
-    "phased_geno_prob",
-]
-ROWS_PER_INDIVIDUAL = {
-    "dosage": 1,
-    "geno_0.333": 1,
-    "hap_0.5": 2,
-    "geno_prob": 3,
-    "phased_geno_prob": 4,
-    "seg_prob": 4,
-}
-SEG_PROB_START_GEN = 2
-HAP_FILE = "hap_0.5"
-TINYPEEL_DIRECT_IS_WARMED_UP = False
-DEFAULT_VISUALIZATION_METRICS = ("marker_corr", "abs_diff", "correct_rate")
-ACCURACY_REPORT_VALUE_NAMES = (
-    "population",
-    "generation_1",
-    "generation_2",
-    "generation_3",
-    "generation_4",
-    "generation_5",
-)
-
-
-def get_params():
-    with open(SIMULATION_PARAMETERS_FILE, "r") as file:
-        sim_params = [line.strip().split() for line in file]
-
-    params = {}
-    for param_name, param_value in sim_params:
-        params[param_name] = float(param_value)
-
-    return params
-
-
-def sim_path():
-    """Return the simulation directory used by the accuracy tests.
-
-    :return: Relative path to the simulation fixture directory.
-    :rtype: str
-    """
-
-    return SIMULATION_FIXTURE_DIR
-
-
-def get_accuracy_benchmark_output_root(run_name="test_accu"):
-    """Return the root directory for direct-call accuracy benchmark outputs."""
-
-    return os.path.join(ACCURACY_TEST_ROOT, f"outputs_{run_name}")
-
-
-def get_accuracy_benchmark_report_root(run_name="test_accu"):
-    """Return the root directory for direct-call accuracy benchmark reports."""
-
-    return os.path.join(ACCURACY_TEST_ROOT, f"reports_{run_name}")
-
-
-def prepare_directory(path):
-    """Create an empty directory at ``path``."""
-
-    if os.path.exists(path):
-        shutil.rmtree(path)
-
-    os.makedirs(path, exist_ok=True)
-
-
-def ensure_directory(path):
-    """Create ``path`` if it does not exist."""
-
-    os.makedirs(path, exist_ok=True)
-
-
-def build_accuracy_output_path(name, run_name="test_accu"):
-    """Build the per-case direct-call benchmark output directory path."""
-
-    return os.path.join(get_accuracy_benchmark_output_root(run_name), name)
-
-
-def build_accuracy_report_path(run_name="test_accu"):
-    """Build the direct-call benchmark report file path."""
-
-    return os.path.join(get_accuracy_benchmark_report_root(run_name), "accu_report.txt")
-
-
-def run_command(command):
-    use_shell = isinstance(command, str)
-    result = subprocess.run(
-        command,
-        shell=use_shell,
-        text=True,
-        capture_output=True,
-    )
-
-    if result.returncode != 0:
-        raise RuntimeError(
-            "AlphaPeel run failed\n"
-            f"Command: {command}\n"
-            f"Exit code: {result.returncode}\n"
-            f"STDOUT:\n{result.stdout}\n"
-            f"STDERR:\n{result.stderr}"
-        )
-
-    return result
-
-
-def get_accuracy_benchmark_cases():
-    """Return the full accuracy benchmark case matrix."""
-
-    return [
-        ("single", None, None, None, None, None, None, None, None),
-        (
-            "single",
-            "est_start_alt_allele_prob",
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-        ),
-        ("single", None, None, None, None, None, "est_alt_allele_prob", None, None),
-        ("multi", None, None, None, None, None, None, None, None),
-        (
-            "multi",
-            "est_start_alt_allele_prob",
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-        ),
-        ("multi", None, None, None, None, None, "est_alt_allele_prob", None, None),
-        (
-            "multi",
-            "est_start_alt_allele_prob",
-            "est_geno_error_prob",
-            "est_seq_error_prob",
-            None,
-            None,
-            None,
-            None,
-            None,
-        ),
-        ("multi", None, None, None, "seq_file", None, None, None, None),
-        (
-            "multi",
-            "est_start_alt_allele_prob",
-            None,
-            None,
-            "seq_file",
-            None,
-            None,
-            None,
-            None,
-        ),
-        (
-            "multi",
-            "est_start_alt_allele_prob",
-            "est_geno_error_prob",
-            "est_seq_error_prob",
-            "seq_file",
-            None,
-            None,
-            None,
-            None,
-        ),
-        ("hybrid", None, None, None, None, None, None, None, None),
-        ("hybrid", None, None, None, "seq_file", None, None, None, None),
-        (
-            "single",
-            None,
-            None,
-            None,
-            None,
-            "alt_allele_prob_file",
-            None,
-            "metafounder",
-            None,
-        ),
-        (
-            "single",
-            "est_start_alt_allele_prob",
-            None,
-            None,
-            None,
-            None,
-            None,
-            "metafounder",
-            None,
-        ),
-        (
-            "single",
-            None,
-            None,
-            None,
-            None,
-            None,
-            "est_alt_allele_prob",
-            "metafounder",
-            None,
-        ),
-        (
-            "single",
-            "est_start_alt_allele_prob",
-            None,
-            None,
-            None,
-            None,
-            "est_alt_allele_prob",
-            "metafounder",
-            None,
-        ),
-        (
-            "single",
-            None,
-            None,
-            None,
-            None,
-            "alt_allele_prob_file",
-            "est_alt_allele_prob",
-            "metafounder",
-            None,
-        ),
-        ("single", None, None, None, None, None, None, None, "x_chr"),
-        ("multi", None, None, None, None, None, None, None, "x_chr"),
-        ("hybrid", None, None, None, None, None, None, None, "x_chr"),
-    ]
-
-
-def build_accuracy_case_name(
-    method,
-    est_start_alt_allele_prob,
-    est_geno_error_prob,
-    est_seq_error_prob,
-    seq_file,
-    alt_allele_prob_file,
-    est_alt_allele_prob,
-    metafounder,
-    x_chr,
-):
-    """Build the benchmark case name from the accuracy case parameters."""
-
-    return "_".join(
-        str(param)
-        for param in [
-            method,
-            est_start_alt_allele_prob,
-            est_geno_error_prob,
-            est_seq_error_prob,
-            seq_file,
-            alt_allele_prob_file,
-            est_alt_allele_prob,
-            metafounder,
-            x_chr,
-        ]
-        if param
-    )
-
-
-def _alpha_peel_arguments(method):
-    """Return the common AlphaPeel arguments for an accuracy run."""
-
-    return {
-        "method": method,
-        **DEFAULT_ALPHA_PEEL_ARGS,
-    }
-
-
-def _input_files_for_case(method, seq_file=False, alt_allele_prob_file=False):
-    """Return the fixture input file keys needed for an accuracy case."""
-
-    input_files = ["ped_file", "seq_file" if seq_file else "geno_file"]
-
-    if alt_allele_prob_file:
-        input_files.append("alt_allele_prob_file")
-    if method == "hybrid":
-        input_files.extend(["map_file", "seg_map_file", "seg_file"])
-
-    return input_files
-
-
-def _fixture_file_prefix(metafounder=False, x_chr=False):
-    """Return the fixture filename prefix for special benchmark modes."""
-
-    if metafounder:
-        return "metafounder_"
-    if x_chr:
-        return "X_chr_"
-    return ""
-
-
-def _fixture_file_path(sim_path, file_name, metafounder=False, x_chr=False):
-    """Build the path to one benchmark fixture file."""
-
-    prefix = _fixture_file_prefix(metafounder=metafounder, x_chr=x_chr)
-    return os.path.join(sim_path, f"{prefix}{file_name}.txt")
-
-
-def _input_file_path(
-    sim_path,
-    file_name,
-    metafounder=False,
-    x_chr=False,
-    file_overrides=None,
-):
-    """Build an input file path, allowing a case to override selected files."""
-
-    if file_overrides and file_name in file_overrides:
-        return file_overrides[file_name]
-
-    return _fixture_file_path(
-        sim_path,
-        file_name,
-        metafounder=metafounder,
-        x_chr=x_chr,
-    )
-
-
-def _add_argument(argv, key, value):
-    """Append a CLI argument to ``argv`` when building a direct-call run."""
-
-    argv.append(f"-{key}")
-    if value is not None:
-        argv.append(value)
-
-
-def generate_accuracy_argv(
-    sim_path,
-    method,
-    est_start_alt_allele_prob,
-    est_geno_error_prob,
-    est_seq_error_prob,
-    seq_file,
-    alt_allele_prob_file,
-    est_alt_allele_prob,
-    metafounder,
-    x_chr,
-    output_path,
-    file_overrides=None,
-    extra_input_files=None,
-    input_files_method=None,
-):
-    """Generate an ``argv`` list for a direct ``tinypeel.main`` accuracy run."""
-
-    argv = []
-    input_files = _input_files_for_case(
-        input_files_method or method,
-        seq_file,
-        alt_allele_prob_file,
-    )
-    if extra_input_files:
-        input_files.extend(extra_input_files)
-
-    arguments = _alpha_peel_arguments(method)
-
-    if est_start_alt_allele_prob:
-        arguments["est_start_alt_allele_prob"] = None
-    if est_geno_error_prob and est_seq_error_prob:
-        arguments["est_geno_error_prob"] = None
-        arguments["est_seq_error_prob"] = None
-    if est_alt_allele_prob:
-        arguments["est_alt_allele_prob"] = None
-    for file_name in input_files:
-        argv.extend(
-            [
-                f"-{file_name}",
-                _input_file_path(
-                    sim_path,
-                    file_name,
-                    metafounder=metafounder,
-                    x_chr=x_chr,
-                    file_overrides=file_overrides,
-                ),
-            ]
-        )
-
-    if x_chr:
-        argv.append("-x_chr")
-
-    for key, value in arguments.items():
-        _add_argument(argv, key, value)
-
-    argv.extend(["-out_file", f"{output_path}{os.sep}"])
-
-    return argv
-
-
-def generate_command(
-    sim_path,
-    method,
-    output_path,
-    est_start_alt_allele_prob=False,
-    est_geno_error_prob=False,
-    est_seq_error_prob=False,
-    seq_file=False,
-    alt_allele_prob_file=False,
-    est_alt_allele_prob=False,
-    metafounder=False,
-    x_chr=False,
-    file_overrides=None,
-    extra_input_files=None,
-    input_files_method=None,
-):
-    """Generate the command used to run AlphaPeel for a test case.
-
-    :param sim_path: Directory containing simulation input files.
-    :type sim_path: str
-    :param method: AlphaPeel method to run, such as ``single``, ``multi`` or
-        ``hybrid``.
-    :type method: str
-    :param output_path: Output directory for AlphaPeel results.
-    :type output_path: str
-    :return: The command argv to pass to ``subprocess.run``.
-    :rtype: list[str]
-    """
-
-    argv = generate_accuracy_argv(
-        sim_path,
-        method,
-        est_start_alt_allele_prob,
-        est_geno_error_prob,
-        est_seq_error_prob,
-        seq_file,
-        alt_allele_prob_file,
-        est_alt_allele_prob,
-        metafounder,
-        x_chr,
-        output_path,
-        file_overrides=file_overrides,
-        extra_input_files=extra_input_files,
-        input_files_method=input_files_method,
-    )
-
-    return ["AlphaPeel", *argv]
-
-
-def run_tinypeel_direct(argv):
-    """Run AlphaPeel by calling ``src.tinypeel.tinypeel.main`` directly."""
-
-    tinypeel.main(argv=argv)
-
-
-def benchmark_tinypeel_direct(argv, output_path=None, warmup=True):
-    """Warm up JIT compilation, then run AlphaPeel and return elapsed seconds."""
-
-    global TINYPEEL_DIRECT_IS_WARMED_UP
-
-    if warmup and not TINYPEEL_DIRECT_IS_WARMED_UP:
-        run_tinypeel_direct(argv)
-        TINYPEEL_DIRECT_IS_WARMED_UP = True
-        if output_path is not None:
-            prepare_directory(output_path)
-
-    start_time = time.perf_counter()
-    run_tinypeel_direct(argv)
-
-    return time.perf_counter() - start_time
-
-
-def _accuracy_files(method, metafounder=False):
-    """Return output file stems to compare for an accuracy run."""
-
-    file_names = METAFOUNDER_ACCURACY_FILES if metafounder else BASE_ACCURACY_FILES
-    file_names = list(file_names)
-
-    if method == "multi":
-        file_names.append("seg_prob")
-
-    return file_names
-
-
-def _accuracy_dimensions(get_params):
-    """Return generation, individual, and locus dimensions from parameters."""
-
-    n_gen = int(get_params["nGen"])
-    n_ind_per_gen = int(get_params["nInd"] / n_gen)
-    n_loci_all = int(get_params["nLociAll"])
-
-    return n_gen, n_ind_per_gen, n_loci_all
-
-
-def _load_accuracy_matrix(path, n_loci_all):
-    """Load an AlphaPeel output or truth matrix for accuracy comparison."""
-
-    return np.loadtxt(path, usecols=np.arange(1, n_loci_all + 1))
-
-
-def _truth_file_path(sim_path, file_name, metafounder=False, x_chr=False):
-    """Build the path to a truth file for an accuracy comparison."""
-
-    if metafounder:
-        true_file = f"true-metafounder_{file_name}.txt"
-    elif x_chr:
-        true_file = f"true-X_chr_{file_name}.txt"
-    else:
-        true_file = f"true-{file_name}.txt"
-
-    return os.path.join(sim_path, true_file)
-
-
-def _comparison_slice(matrix, file_name, n_ind_per_gen, n_row_per_ind):
-    """Return the portion of a matrix used by the test accuracy report."""
-
-    if file_name == "seg_prob":
-        start = SEG_PROB_START_GEN * (n_ind_per_gen * n_row_per_ind)
-        return matrix[start:, 1:]
-
-    return matrix[:, 1:]
-
-
-def _generation_slice(matrix, gen, n_ind_per_gen, n_row_per_ind):
-    """Return rows belonging to one generation."""
-
-    start = gen * (n_ind_per_gen * n_row_per_ind)
-    end = (gen + 1) * (n_ind_per_gen * n_row_per_ind)
-
-    return matrix[start:end]
-
-
-def _load_accuracy_pair(output_path, sim_path, file_name, n_loci_all, **truth_kwargs):
-    """Load one output/truth matrix pair for an accuracy comparison."""
-
-    file_path = os.path.join(output_path, f".{file_name}.txt")
-    true_path = _truth_file_path(sim_path, file_name, **truth_kwargs)
-
-    return (
-        _load_accuracy_matrix(file_path, n_loci_all),
-        _load_accuracy_matrix(true_path, n_loci_all),
-    )
-
-
-def _test_truth_file_path(sim_path, file_name):
-    """Build the truth path used by the pytest accuracy report."""
-
-    return os.path.join(sim_path, f"true-{file_name}.txt")
-
-
-def _load_test_accuracy_pair(output_path, sim_path, file_name, n_loci_all):
-    """Load one output/truth matrix pair for the pytest accuracy report."""
-
-    file_path = os.path.join(output_path, f".{file_name}.txt")
-    true_path = _test_truth_file_path(sim_path, file_name)
-
-    try:
-        output = _load_accuracy_matrix(file_path, n_loci_all)
-    except ValueError:
-        print(f"Error loading {file_path}")
-        return None
-
-    try:
-        truth = _load_accuracy_matrix(true_path, n_loci_all)
-    except ValueError:
-        print(f"Error loading {true_path}")
-        return None
-
-    return output, truth
-
-
-def _mask_x_chr_hap_missing(output, truth, file_name, x_chr):
-    """Mask missing X chromosome haplotype truth values before comparison."""
-
-    if x_chr and file_name == "hap_0.5":
-        output[truth == 9] = 0
-        truth[truth == 9] = 0
-
-
-def _overall_comparison(output, truth, file_name, n_ind_per_gen, n_row_per_ind):
-    """Return output/truth slices used for whole-file metrics."""
-
-    return (
-        _comparison_slice(output, file_name, n_ind_per_gen, n_row_per_ind),
-        _comparison_slice(truth, file_name, n_ind_per_gen, n_row_per_ind),
-    )
-
-
-def _ind_corr_for_file(output, truth, file_name, n_ind_per_gen, n_row_per_ind):
-    """Return individual correlation with seg_prob generation handling."""
-
-    if file_name == "seg_prob":
-        return get_ind_corr(
-            output,
-            truth,
-            n_ind_per_gen,
-            n_row_per_ind,
-            start_gen=SEG_PROB_START_GEN,
-        )
-
-    return get_ind_corr(output, truth, n_ind_per_gen, n_row_per_ind)
-
-
-def _write_accuracy_metric(file_out, file_name, label, metric_name, value):
-    """Write one metric line to an accuracy report."""
-
-    file_out.write(f"{file_name},{label},{metric_name},{value}\n")
-
-
-def _seg_prob_generation_is_skipped(file_name, gen):
-    """Return whether a per-generation seg_prob metric should be reported as nan."""
-
-    return file_name == "seg_prob" and gen < SEG_PROB_START_GEN
-
-
-def _per_generation_metric(
-    metric_func,
-    output,
-    truth,
-    file_name,
-    n_gen,
-    n_ind_per_gen,
-    n_row_per_ind,
-    *metric_args,
-):
-    """Return a list of per-generation metric values as strings."""
-
-    values = []
-    for gen in range(n_gen):
-        if _seg_prob_generation_is_skipped(file_name, gen):
-            values.append("nan")
-            continue
-
-        values.append(
-            str(
-                metric_func(
-                    _generation_slice(output, gen, n_ind_per_gen, n_row_per_ind),
-                    _generation_slice(truth, gen, n_ind_per_gen, n_row_per_ind),
-                    *metric_args,
-                )
-            )
-        )
-
-    return values
-
-
-def _benchmark_metric_values(
-    metric_func,
-    output,
-    truth,
-    file_name,
-    n_gen,
-    n_ind_per_gen,
-    n_row_per_ind,
-    *metric_args,
-):
-    """Return whole-file and per-generation benchmark metric values."""
-
-    overall_output, overall_truth = _overall_comparison(
-        output, truth, file_name, n_ind_per_gen, n_row_per_ind
-    )
-    overall_value = metric_func(overall_output, overall_truth, *metric_args)
-
-    return [str(overall_value)] + _per_generation_metric(
-        metric_func,
-        output,
-        truth,
-        file_name,
-        n_gen,
-        n_ind_per_gen,
-        n_row_per_ind,
-        *metric_args,
-    )
-
-
-def _benchmark_ind_corr_values(
-    output,
-    truth,
-    file_name,
-    n_gen,
-    n_ind_per_gen,
-    n_row_per_ind,
-):
-    """Return whole-file and per-generation individual-correlation values."""
-
-    overall_start_gen = None
-    if file_name == "seg_prob":
-        overall_start_gen = SEG_PROB_START_GEN
-
-    values = [
-        str(
-            get_ind_corr(
-                output,
-                truth,
-                n_ind_per_gen,
-                n_row_per_ind,
-                start_gen=overall_start_gen,
-            )
-        )
-    ]
-
-    return values + _per_generation_metric(
-        get_ind_corr,
-        output,
-        truth,
-        file_name,
-        n_gen,
-        n_ind_per_gen,
-        n_row_per_ind,
-        n_ind_per_gen,
-        n_row_per_ind,
-    )
-
-
-def _safe_rate(numerator, denominator):
-    """Return a rate, or nan when the denominator is empty."""
-
-    if denominator == 0:
-        return np.nan
-
-    return numerator / denominator
-
-
-def get_hap_switch_error_metrics(called_file, true_file, n_ind, n_loci_all):
-    """Calculate switch, phase, and related haplotype call metrics."""
-
-    switch_error_count = 0
-    phase_error_count = 0
-    uncalled_count = 0
-    wrong_homo_count = 0
-    true_hetero_count = 0
-    homo_count = 0
-    hetero_count = 0
-
-    for ind in range(n_ind):
-        hap_p_new = called_file[ind * 2]
-        hap_m_new = called_file[ind * 2 + 1]
-        hap_p_true = true_file[ind * 2]
-        hap_m_true = true_file[ind * 2 + 1]
-
-        switched = False
-        for loci in range(n_loci_all):
-            if hap_p_true[loci] == hap_m_true[loci]:
-                homo_count += 1
-            else:
-                hetero_count += 1
-
-            if hap_p_new[loci] == 9 or hap_m_new[loci] == 9:
-                uncalled_count += 1
-                continue
-
-            if hap_p_true[loci] + hap_m_true[loci] == 1:
-                if hap_p_new[loci] == hap_m_new[loci]:
-                    wrong_homo_count += 1
-                    continue
-
-                true_hetero_count += 1
-                if hap_p_new[loci] != hap_p_true[loci]:
-                    phase_error_count += 1
-
-                if (hap_p_new[loci] != hap_p_true[loci] and switched is False) or (
-                    hap_p_new[loci] == hap_p_true[loci] and switched is True
-                ):
-                    switched = not switched
-                    switch_error_count += 1
-
-    genotype_count = n_ind * n_loci_all
-    switch_opportunity_count = n_ind * (n_loci_all - 1)
-
-    return [
-        (
-            "switch_error_rate",
-            _safe_rate(switch_error_count, switch_opportunity_count),
-        ),
-        ("phase_error_rate", _safe_rate(phase_error_count, genotype_count)),
-        ("uncalled_rate", _safe_rate(uncalled_count, genotype_count)),
-        ("wrong_homozygote_rate", _safe_rate(wrong_homo_count, genotype_count)),
-        (
-            "correct_heterozygote_rate",
-            _safe_rate(true_hetero_count, genotype_count),
-        ),
-        ("homozygote_count", homo_count),
-        ("heterozygote_count", hetero_count),
-        ("homo_to_hetero_ratio", _safe_rate(homo_count, hetero_count)),
-    ]
-
-
-def _benchmark_hap_switch_error_metrics(output, truth, n_gen, n_ind_per_gen):
-    """Return whole-file and per-generation switch-error metrics."""
-
-    n_loci_all = output.shape[1]
-    metric_values = [
-        (metric_name, [str(value)])
-        for metric_name, value in get_hap_switch_error_metrics(
-            output,
-            truth,
-            n_gen * n_ind_per_gen,
-            n_loci_all,
-        )
-    ]
-    values_by_metric = dict(metric_values)
-
-    for gen in range(n_gen):
-        gen_output = _generation_slice(
-            output,
-            gen,
-            n_ind_per_gen,
-            ROWS_PER_INDIVIDUAL[HAP_FILE],
-        )
-        gen_truth = _generation_slice(
-            truth,
-            gen,
-            n_ind_per_gen,
-            ROWS_PER_INDIVIDUAL[HAP_FILE],
-        )
-
-        for metric_name, value in get_hap_switch_error_metrics(
-            gen_output,
-            gen_truth,
-            n_ind_per_gen,
-            n_loci_all,
-        ):
-            values_by_metric[metric_name].append(str(value))
-
-    return [
-        (metric_name, values_by_metric[metric_name]) for metric_name, _ in metric_values
-    ]
-
-
-def _write_test_hap_switch_error_metrics(
-    file_out,
-    file_name,
-    method,
-    output,
-    truth,
-    n_gen,
-    n_ind_per_gen,
-):
-    """Write test accuracy switch-error metrics for haplotype output."""
-
-    for metric_name, value in get_hap_switch_error_metrics(
-        output,
-        truth,
-        n_gen * n_ind_per_gen,
-        output.shape[1],
-    ):
-        if metric_name in [
-            "homozygote_count",
-            "heterozygote_count",
-            "homo_to_hetero_ratio",
-        ]:
-            continue
-        _write_accuracy_metric(file_out, file_name, method, metric_name, value)
-
-
-def _write_benchmark_hap_switch_error_metrics(
-    file_out,
-    file_name,
-    name,
-    output,
-    truth,
-    n_gen,
-    n_ind_per_gen,
-):
-    """Write benchmark switch-error metrics for haplotype output."""
-
-    for metric_name, values in _benchmark_hap_switch_error_metrics(
-        output,
-        truth,
-        n_gen,
-        n_ind_per_gen,
-    ):
-        _write_accuracy_metric(file_out, file_name, name, metric_name, values)
-
-
-def assess_test_accuracy(
-    sim_path,
-    get_params,
-    output_path,
-    method,
-    file_out,
-):
-    file_to_check = _accuracy_files(method)
-    n_gen, n_ind_per_gen, n_loci_all = _accuracy_dimensions(get_params)
-
-    for file_name in file_to_check:
-        n_row_per_ind = ROWS_PER_INDIVIDUAL[file_name]
-        accuracy_pair = _load_test_accuracy_pair(
-            output_path, sim_path, file_name, n_loci_all
-        )
-        if accuracy_pair is None:
-            continue
-
-        new_file, true_file = accuracy_pair
-        comparison_output, comparison_truth = _overall_comparison(
-            new_file, true_file, file_name, n_ind_per_gen, n_row_per_ind
-        )
-        marker_corr = get_marker_corr(
-            comparison_output,
-            comparison_truth,
-        )
-        _write_accuracy_metric(file_out, file_name, method, "marker_corr", marker_corr)
-
-        ind_corr = _ind_corr_for_file(
-            new_file,
-            true_file,
-            file_name,
-            n_ind_per_gen,
-            n_row_per_ind,
-        )
-        _write_accuracy_metric(file_out, file_name, method, "ind_corr", ind_corr)
-
-        abs_diff = get_abs_diff(
-            comparison_output,
-            comparison_truth,
-            n_row_per_ind,
-        )
-        _write_accuracy_metric(file_out, file_name, method, "abs_diff", abs_diff)
-
-        if file_name == "hap_0.5":
-            _write_test_hap_switch_error_metrics(
-                file_out,
-                file_name,
-                method,
-                new_file,
-                true_file,
-                n_gen,
-                n_ind_per_gen,
-            )
-
-        if file_name == "seg_prob":
-            correct_rate = get_correct_rate(comparison_output, comparison_truth)
-            _write_accuracy_metric(
-                file_out, file_name, method, "correct_rate", correct_rate
-            )
-
-
-def assess_accuracy(
-    sim_path,
-    get_params,
-    output_path,
-    name,
-    method,
-    file_out,
-    metafounder,
-    x_chr,
-):
-    """Assess output accuracy against truth files."""
-
-    file_to_check = _accuracy_files(method, metafounder)
-    n_gen, n_ind_per_gen, n_loci_all = _accuracy_dimensions(get_params)
-
-    print(" ")
-    print(f"Test: {name}")
-
-    for file_name in file_to_check:
-        n_row_per_ind = ROWS_PER_INDIVIDUAL[file_name]
-        new_file, true_file = _load_accuracy_pair(
-            output_path,
-            sim_path,
-            file_name,
-            n_loci_all,
-            metafounder=metafounder,
-            x_chr=x_chr,
-        )
-        _mask_x_chr_hap_missing(new_file, true_file, file_name, x_chr)
-
-        marker_corr = _benchmark_metric_values(
-            get_marker_corr,
-            new_file,
-            true_file,
-            file_name,
-            n_gen,
-            n_ind_per_gen,
-            n_row_per_ind,
-        )
-        _write_accuracy_metric(file_out, file_name, name, "marker_corr", marker_corr)
-
-        ind_corr = _benchmark_ind_corr_values(
-            new_file,
-            true_file,
-            file_name,
-            n_gen,
-            n_ind_per_gen,
-            n_row_per_ind,
-        )
-        _write_accuracy_metric(file_out, file_name, name, "ind_corr", ind_corr)
-
-        abs_diff = _benchmark_metric_values(
-            get_abs_diff,
-            new_file,
-            true_file,
-            file_name,
-            n_gen,
-            n_ind_per_gen,
-            n_row_per_ind,
-            n_row_per_ind,
-        )
-        _write_accuracy_metric(file_out, file_name, name, "abs_diff", abs_diff)
-
-        if file_name == "hap_0.5":
-            _write_benchmark_hap_switch_error_metrics(
-                file_out,
-                file_name,
-                name,
-                new_file,
-                true_file,
-                n_gen,
-                n_ind_per_gen,
-            )
-
-        if file_name == "seg_prob":
-            correct_rate = _benchmark_metric_values(
-                get_correct_rate,
-                new_file,
-                true_file,
-                file_name,
-                n_gen,
-                n_ind_per_gen,
-                n_row_per_ind,
-            )
-            _write_accuracy_metric(
-                file_out, file_name, name, "correct_rate", correct_rate
-            )
-
-
-def _read_map_marker_names(path):
-    """Return marker names from an AlphaPeel map file."""
-
-    with open(path, "r") as file:
-        return [line.split()[1] for line in file if line.strip()]
-
-
-def _subset_locus_file(source_path, target_path, columns):
-    """Write a genotype or sequence file containing only selected loci."""
-
-    data = np.loadtxt(source_path, dtype=np.int64)
-    np.savetxt(target_path, data[:, columns], fmt="%d")
-
-
-def _hybrid_multi_output_path(output_path):
-    """Return the first-stage multi-locus output directory for a hybrid case."""
-
-    return os.path.join(output_path, "multi_stage")
-
-
-def _hybrid_subset_input_dir(output_path):
-    """Return the directory for first-stage subset genotype or sequence inputs."""
-
-    return os.path.join(output_path, "subset_inputs")
-
-
-def _prepare_hybrid_multi_inputs(
-    sim_path,
-    seq_file,
-    metafounder,
-    x_chr,
-    output_path,
-):
-    """Create subset inputs for the first stage of a hybrid benchmark."""
-
-    subset_input_dir = _hybrid_subset_input_dir(output_path)
-    ensure_directory(subset_input_dir)
-
-    map_path = _fixture_file_path(
-        sim_path,
-        "map_file",
-        metafounder=metafounder,
-        x_chr=x_chr,
-    )
-    seg_map_path = _fixture_file_path(
-        sim_path,
-        "seg_map_file",
-        metafounder=metafounder,
-        x_chr=x_chr,
-    )
-    full_markers = _read_map_marker_names(map_path)
-    subset_markers = _read_map_marker_names(seg_map_path)
-    marker_columns = {marker: index + 1 for index, marker in enumerate(full_markers)}
-    missing_markers = [
-        marker for marker in subset_markers if marker not in marker_columns
-    ]
-    if missing_markers:
-        raise ValueError(
-            "Hybrid seg_map_file contains markers that are not in map_file: "
-            f"{missing_markers}"
-        )
-
-    columns = [0] + [marker_columns[marker] for marker in subset_markers]
-
-    locus_file_name = "seq_file" if seq_file else "geno_file"
-    source_path = _fixture_file_path(
-        sim_path,
-        locus_file_name,
-        metafounder=metafounder,
-        x_chr=x_chr,
-    )
-    subset_path = os.path.join(subset_input_dir, f"{locus_file_name}.txt")
-    _subset_locus_file(source_path, subset_path, columns)
-
-    return {
-        locus_file_name: subset_path,
-        "map_file": seg_map_path,
-    }
-
-
-def _run_commands(commands):
-    """Run a sequence of shell commands."""
-
-    for command in commands:
-        run_command(command)
-
-
-def _run_tinypeel_direct_sequence(argvs):
-    """Run a sequence of direct tinypeel calls."""
-
-    for argv in argvs:
-        run_tinypeel_direct(argv)
-
-
-def run_accuracy_case(
-    get_params,
-    sim_path,
-    method,
-    est_start_alt_allele_prob=False,
-    est_geno_error_prob=False,
-    est_seq_error_prob=False,
-    seq_file=False,
-    alt_allele_prob_file=False,
-    est_alt_allele_prob=False,
-    metafounder=False,
-    x_chr=False,
-    benchmark=None,
-    run_name="test_accu",
-):
-    """Run AlphaPeel and evaluate outputs based on the specified parameters."""
-
-    name = build_accuracy_case_name(
-        method,
-        est_start_alt_allele_prob,
-        est_geno_error_prob,
-        est_seq_error_prob,
-        seq_file,
-        alt_allele_prob_file,
-        est_alt_allele_prob,
-        metafounder,
-        x_chr,
-    )
-    output_path = build_accuracy_output_path(name, run_name)
-    prepare_directory(output_path)
-    runtime_seconds = None
-
-    if method == "hybrid":
-        multi_output_path = _hybrid_multi_output_path(output_path)
-        prepare_directory(multi_output_path)
-        multi_file_overrides = _prepare_hybrid_multi_inputs(
-            sim_path,
-            seq_file,
-            metafounder,
-            x_chr,
-            output_path,
-        )
-        hybrid_file_overrides = {
-            "seg_file": os.path.join(multi_output_path, ".seg_prob.txt")
-        }
-
-        if run_name == "test_accu":
-            commands = [
-                generate_command(
-                    sim_path,
-                    "multi",
-                    multi_output_path,
-                    est_start_alt_allele_prob,
-                    est_geno_error_prob,
-                    est_seq_error_prob,
-                    seq_file,
-                    alt_allele_prob_file,
-                    est_alt_allele_prob,
-                    metafounder,
-                    x_chr,
-                    file_overrides=multi_file_overrides,
-                    extra_input_files=["map_file"],
-                ),
-                generate_command(
-                    sim_path,
-                    "single",
-                    output_path,
-                    est_start_alt_allele_prob,
-                    est_geno_error_prob,
-                    est_seq_error_prob,
-                    seq_file,
-                    alt_allele_prob_file,
-                    est_alt_allele_prob,
-                    metafounder,
-                    x_chr,
-                    file_overrides=hybrid_file_overrides,
-                    input_files_method="hybrid",
-                ),
-            ]
-            benchmark(_run_commands, commands)
-        else:
-            argvs = [
-                generate_accuracy_argv(
-                    sim_path,
-                    "multi",
-                    est_start_alt_allele_prob,
-                    est_geno_error_prob,
-                    est_seq_error_prob,
-                    seq_file,
-                    alt_allele_prob_file,
-                    est_alt_allele_prob,
-                    metafounder,
-                    x_chr,
-                    multi_output_path,
-                    file_overrides=multi_file_overrides,
-                    extra_input_files=["map_file"],
-                ),
-                generate_accuracy_argv(
-                    sim_path,
-                    "single",
-                    est_start_alt_allele_prob,
-                    est_geno_error_prob,
-                    est_seq_error_prob,
-                    seq_file,
-                    alt_allele_prob_file,
-                    est_alt_allele_prob,
-                    metafounder,
-                    x_chr,
-                    output_path,
-                    file_overrides=hybrid_file_overrides,
-                    input_files_method="hybrid",
-                ),
-            ]
-            _run_tinypeel_direct_sequence(argvs)
-
-    elif run_name == "test_accu":
-        command = generate_command(
-            sim_path,
-            method,
-            output_path,
-            est_start_alt_allele_prob,
-            est_geno_error_prob,
-            est_seq_error_prob,
-            seq_file,
-            alt_allele_prob_file,
-            est_alt_allele_prob,
-            metafounder,
-            x_chr,
-        )
-        benchmark(run_command, command)
-
-    else:
-        argv = generate_accuracy_argv(
-            sim_path,
-            method,
-            est_start_alt_allele_prob,
-            est_geno_error_prob,
-            est_seq_error_prob,
-            seq_file,
-            alt_allele_prob_file,
-            est_alt_allele_prob,
-            metafounder,
-            x_chr,
-            output_path,
-        )
-        runtime_seconds = benchmark_tinypeel_direct(argv, output_path=output_path)
-
-    report_path = build_accuracy_report_path(run_name)
-    ensure_directory(os.path.dirname(report_path))
-    with open(report_path, "a") as file_out:
-        if run_name == "test_accu":
-            assess_test_accuracy(
-                sim_path,
-                get_params,
-                output_path,
-                method,
-                file_out,
-            )
-        elif run_name == "benchmark":
-            if runtime_seconds is not None:
-                _write_accuracy_metric(
-                    file_out,
-                    "runtime",
-                    name,
-                    "elapsed_seconds",
-                    runtime_seconds,
-                )
-
-            assess_accuracy(
-                sim_path,
-                get_params,
-                output_path,
-                name,
-                method,
-                file_out,
-                metafounder,
-                x_chr,
-            )
-
-
-def run_full_accuracy_suite(run_name="benchmark"):
-    """Run the full direct-call accuracy benchmark suite."""
-
-    prepare_directory(get_accuracy_benchmark_output_root(run_name))
-    prepare_directory(get_accuracy_benchmark_report_root(run_name))
-
-    for case in get_accuracy_benchmark_cases():
-        run_accuracy_case(
-            get_params(),
-            sim_path(),
-            *case,
-            benchmark=None,
-            run_name=run_name,
-        )
-
-
-def get_marker_corr(output, real):
-    """Compute the average marker-wise Pearson correlation coefficient between two arrays.
-
-    :param output: Output matrix to compare.
-    :type output: numpy.ndarray
-    :param real: Reference matrix to compare against.
-    :type real: numpy.ndarray
-    :return: Rounded mean marker correlation coefficient.
-    :rtype: float
-    """
-
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore")
-        accus = np.array(
-            [np.corrcoef(real[:, i], output[:, i])[0, 1] for i in range(real.shape[1])]
-        )
-        return round(np.nanmean(accus), 4)
-
-
-def get_ind_corr(output, real, nIndPerGen, n_row_per_ind, start_gen=None, end_gen=None):
-    """Compute the average individual-wise Pearson correlation coefficient between two arrays.
-
-    :param output: Output matrix to compare.
-    :type output: numpy.ndarray
-    :param real: Reference matrix to compare against.
-    :type real: numpy.ndarray
-    :param nIndPerGen: Number of individuals per generation.
-    :type nIndPerGen: int
-    :param n_row_per_ind: Number of rows per individual in the output.
-    :type n_row_per_ind: int or None
-    :param start_gen: Optional starting generation index used to slice the correlation vector.
-    :type start_gen: int or None
-    :param end_gen: Optional ending generation index used to slice the correlation vector.
-    :type end_gen: int or None
-    :return: Rounded mean individual accuracy.
-    :rtype: float
-    """
-
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore")
-        accus = np.array(
-            [np.corrcoef(real[i, :], output[i, :])[0, 1] for i in range(real.shape[0])]
-        )
-        if isinstance(start_gen, int):
-            if not isinstance(end_gen, int):
-                accus = accus[start_gen * (nIndPerGen * n_row_per_ind) :]
-            else:
-                accus = accus[
-                    start_gen
-                    * (nIndPerGen * n_row_per_ind) : end_gen
-                    * (nIndPerGen * n_row_per_ind)
-                ]
-
-        return round(np.nanmean(accus), 4)
-
-
-def get_abs_diff(output, real, n_row_per_ind):
-    """Sum of absolute difference divided by the sum of the number of loci being counted"""
-    return n_row_per_ind * np.sum(np.abs(output - real)) / real.size
-
-
-def get_correct_rate(output, real):
-    """
-    Summing up the probabilities of the true state from the output data
-    divided by the number of loci being counted
-
-    :param output: Output data
-    :type output: ndarray
-    :param real: Real simulated data
-    :type real: ndarray
-    """
-    return np.sum(output[real == 1]) / (np.size(real) / 4)
-
-
-def _parse_report_float(value):
-    """Parse numeric report values, preserving nan as numpy.nan."""
-
-    if isinstance(value, str) and value.lower() == "nan":
-        return np.nan
-
-    return float(value)
-
-
-def _parse_accuracy_report_value(value):
-    """Parse a scalar or list-valued accuracy report field."""
-
-    value = value.strip()
-    if value.startswith("["):
-        return [_parse_report_float(item) for item in ast.literal_eval(value)]
-
-    return [_parse_report_float(value)]
-
-
-def load_accuracy_report(report_path):
-    """Load an accuracy report into a list of structured records.
-
-    The report uses four comma-separated fields, but list-valued metric fields
-    also contain commas. Splitting each line at the first three commas preserves
-    the current report format without requiring quoted CSV output.
-    """
-
-    records = []
-    with open(report_path, "r") as file:
-        for line_number, line in enumerate(file, start=1):
-            line = line.strip()
-            if not line:
-                continue
-
-            fields = line.split(",", 3)
-            if len(fields) != 4:
-                raise ValueError(
-                    f"Expected 4 fields at {report_path}:{line_number}, "
-                    f"found {len(fields)}."
-                )
-
-            file_name, label, metric_name, raw_value = fields
-            values = _parse_accuracy_report_value(raw_value)
-            record = {
-                "file_name": file_name,
-                "label": label,
-                "metric_name": metric_name,
-                "values": values,
-                "population": values[0],
-            }
-            for index, value_name in enumerate(ACCURACY_REPORT_VALUE_NAMES):
-                if index < len(values):
-                    record[value_name] = values[index]
-
-            records.append(record)
-
-    return records
-
-
-def _matches_optional_filter(value, allowed_values):
-    """Return whether a value passes an optional allow-list filter."""
-
-    return allowed_values is None or value in allowed_values
-
-
-def filter_accuracy_records(
-    records,
-    metric_names=DEFAULT_VISUALIZATION_METRICS,
-    file_names=None,
-    labels=None,
-):
-    """Filter loaded accuracy records for visualization."""
-
-    metric_names = set(metric_names) if metric_names is not None else None
-    file_names = set(file_names) if file_names is not None else None
-    labels = set(labels) if labels is not None else None
-
-    return [
-        record
-        for record in records
-        if _matches_optional_filter(record["metric_name"], metric_names)
-        and _matches_optional_filter(record["file_name"], file_names)
-        and _matches_optional_filter(record["label"], labels)
-    ]
-
-
-def _unique_in_order(values):
-    """Return unique values while preserving first-seen order."""
-
-    unique_values = []
-    seen = set()
-    for value in values:
-        if value not in seen:
-            unique_values.append(value)
-            seen.add(value)
-
-    return unique_values
-
-
-def _safe_plot_name(*parts):
-    """Return a filesystem-friendly plot filename stem."""
-
-    return "_".join(
-        "".join(char if char.isalnum() else "_" for char in str(part)).strip("_")
-        for part in parts
-        if part
-    )
-
-
-def _get_pyplot():
-    """Import matplotlib lazily for report visualization helpers."""
-
-    try:
-        import matplotlib.pyplot as plt
-    except ImportError as error:
-        raise ImportError(
-            "matplotlib is required for accuracy report visualizations."
-        ) from error
-
-    return plt
-
-
-def _plot_output_path(output_dir, *name_parts):
-    """Build an output path for a generated visualization."""
-
-    ensure_directory(output_dir)
-    return os.path.join(output_dir, f"{_safe_plot_name(*name_parts)}.png")
-
-
-def _plot_colors(plt, count):
-    """Return a distinct color sequence for a plot."""
-
-    if count <= 0:
-        return []
-
-    color_map_name = "tab20" if count <= 20 else "nipy_spectral"
-    color_map = plt.get_cmap(color_map_name, count)
-    return [color_map(index) for index in range(count)]
-
-
-def plot_accuracy_population_heatmaps(
-    report_path,
-    output_dir,
-    metric_names=DEFAULT_VISUALIZATION_METRICS,
-    file_names=None,
-    labels=None,
-):
-    """Create one population-level heatmap per selected metric."""
-
-    plt = _get_pyplot()
-    records = filter_accuracy_records(
-        load_accuracy_report(report_path),
-        metric_names=metric_names,
-        file_names=file_names,
-        labels=labels,
-    )
-    generated_paths = []
-
-    for metric_name in _unique_in_order(record["metric_name"] for record in records):
-        metric_records = [
-            record for record in records if record["metric_name"] == metric_name
-        ]
-        metric_labels = _unique_in_order(record["label"] for record in metric_records)
-        metric_files = _unique_in_order(
-            record["file_name"] for record in metric_records
-        )
-        values = {
-            (record["label"], record["file_name"]): record["population"]
-            for record in metric_records
-        }
-        matrix = np.array(
-            [
-                [values.get((label, file_name), np.nan) for file_name in metric_files]
-                for label in metric_labels
-            ],
-            dtype=float,
-        )
-
-        max_label_length = max(len(label) for label in metric_labels)
-        figure_width = max(10, len(metric_files) * 1.4 + max_label_length * 0.12)
-        figure_height = max(4.5, len(metric_labels) * 0.35)
-        figure, axis = plt.subplots(
-            figsize=(figure_width, figure_height),
-            constrained_layout=True,
-        )
-        image = axis.imshow(matrix, aspect="auto", cmap="viridis")
-        axis.set_title(f"{metric_name} population metric")
-        axis.set_xlabel("file_name")
-        axis.set_ylabel("label")
-        axis.set_xticks(np.arange(len(metric_files)))
-        axis.set_xticklabels(metric_files, rotation=45, ha="right")
-        axis.set_yticks(np.arange(len(metric_labels)))
-        axis.set_yticklabels(metric_labels)
-        figure.colorbar(image, ax=axis)
-
-        output_path = _plot_output_path(output_dir, "population_heatmap", metric_name)
-        figure.savefig(output_path, dpi=200)
-        plt.close(figure)
-        generated_paths.append(output_path)
-
-    return generated_paths
-
-
-def plot_accuracy_generation_profiles(
-    report_path,
-    output_dir,
-    metric_names=DEFAULT_VISUALIZATION_METRICS,
-    file_names=None,
-    labels=None,
-):
-    """Create generation profile line plots for selected metrics and files."""
-
-    plt = _get_pyplot()
-    records = filter_accuracy_records(
-        load_accuracy_report(report_path),
-        metric_names=metric_names,
-        file_names=file_names,
-        labels=labels,
-    )
-    generated_paths = []
-
-    metric_file_pairs = _unique_in_order(
-        (record["metric_name"], record["file_name"])
-        for record in records
-        if len(record["values"]) > 1
-    )
-    for metric_name, file_name in metric_file_pairs:
-        plot_records = [
-            record
-            for record in records
-            if record["metric_name"] == metric_name
-            and record["file_name"] == file_name
-            and len(record["values"]) > 1
-        ]
-        if not plot_records:
-            continue
-
-        plot_labels = [record["label"] for record in plot_records]
-        max_label_length = max(len(label) for label in plot_labels)
-        figure_width = min(18, max(10, 6 + max_label_length * 0.12))
-        figure, axis = plt.subplots(figsize=(figure_width, 5.5))
-        generations = np.arange(1, len(ACCURACY_REPORT_VALUE_NAMES))
-        colors = _plot_colors(plt, len(plot_records))
-        for index, record in enumerate(plot_records):
-            generation_values = record["values"][1:]
-            axis.plot(
-                generations,
-                generation_values,
-                color=colors[index],
-                marker="o",
-                linestyle="dashed",
-                label=record["label"],
-            )
-
-        axis.set_title(f"{metric_name} by generation: {file_name}")
-        axis.set_xlabel("generation")
-        axis.set_ylabel(metric_name)
-        axis.set_xticks(generations)
-        axis.legend(loc="center left", bbox_to_anchor=(1, 0.5), fontsize="small")
-        figure.tight_layout()
-
-        output_path = _plot_output_path(
-            output_dir,
-            "generation_profile",
-            metric_name,
-            file_name,
-        )
-        figure.savefig(output_path, dpi=200)
-        plt.close(figure)
-        generated_paths.append(output_path)
-
-    return generated_paths
-
-
-def plot_accuracy_runtime(report_path, output_dir, labels=None):
-    """Create a runtime bar plot for benchmark cases."""
-
-    plt = _get_pyplot()
-    labels = set(labels) if labels is not None else None
-    records = [
-        record
-        for record in load_accuracy_report(report_path)
-        if record["file_name"] == "runtime"
-        and record["metric_name"] == "elapsed_seconds"
-        and _matches_optional_filter(record["label"], labels)
-    ]
-    if not records:
-        return []
-
-    records = sorted(records, key=lambda record: record["population"])
-    runtime_labels = [record["label"] for record in records]
-    runtime_values = [record["population"] for record in records]
-    max_label_length = max(len(label) for label in runtime_labels)
-    figure_width = min(18, max(10, 6 + max_label_length * 0.12))
-    figure_height = max(4.5, len(records) * 0.35)
-
-    figure, axis = plt.subplots(
-        figsize=(figure_width, figure_height),
-        constrained_layout=True,
-    )
-    colors = _plot_colors(plt, len(records))
-    positions = np.arange(len(records))
-    axis.barh(positions, runtime_values, color=colors)
-    axis.set_title("Runtime by benchmark case")
-    axis.set_xlabel("elapsed_seconds")
-    axis.set_ylabel("label")
-    axis.set_yticks(positions)
-    axis.set_yticklabels(runtime_labels)
-    axis.invert_yaxis()
-
-    output_path = _plot_output_path(output_dir, "runtime", "elapsed_seconds")
-    figure.savefig(output_path, dpi=200)
-    plt.close(figure)
-
-    return [output_path]
-
-
-def create_accuracy_report_visualizations(
-    report_path,
-    output_dir,
-    metric_names=DEFAULT_VISUALIZATION_METRICS,
-    file_names=None,
-    labels=None,
-    include_generation_profiles=True,
-    include_runtime=True,
-):
-    """Create the default accuracy report visualization set."""
-
-    generated_paths = plot_accuracy_population_heatmaps(
-        report_path,
-        output_dir,
-        metric_names=metric_names,
-        file_names=file_names,
-        labels=labels,
-    )
-
-    if include_generation_profiles:
-        generated_paths.extend(
-            plot_accuracy_generation_profiles(
-                report_path,
-                output_dir,
-                metric_names=metric_names,
-                file_names=file_names,
-                labels=labels,
-            )
-        )
-
-    if include_runtime:
-        generated_paths.extend(
-            plot_accuracy_runtime(
-                report_path,
-                output_dir,
-                labels=labels,
-            )
-        )
-
-    return generated_paths
-
-
-def calc_switch_error_rate():
-    """
-    Run this function at root directory and input the path to the
-    assessed haplotype file.
-    The SER and PER_intra calculation follows the definition from:
-    https://www.cell.com/hgg-advances/fulltext/S2666-2477(25)00082-X
-    """
-
-    parser = argparse.ArgumentParser(
-        prog="calc_switch_error_rate",
-        description="Calculate switch error rate, phase error rate",
-        epilog="The SER and PER_intra calculation follows the definition from: https://www.cell.com/hgg-advances/fulltext/S2666-2477(25)00082-X",
-    )
-
-    params = get_params()
-
-    nLociAll = int(params["nLociAll"])
-    nInd = int(params["nInd"])
-
-    parser.add_argument(
-        "-true_path",
-        type=str,
-        required=False,
-        default=os.path.join(
-            "tests", "accuracy_tests", "sim_for_alphapeel_accu_test", "true-hap_0.5.txt"
-        ),
-        help="Enter the path of the true haplotype file, default is the simulation path",
-    )
-    parser.add_argument(
-        "-called_path",
-        type=str,
-        required=True,
-        help="Enter the path of the assessed haplotype file",
-    )
-
-    args = parser.parse_args()
-    true_path = args.true_path
-    called_path = args.called_path
-
-    called_file = np.loadtxt(called_path, usecols=np.arange(1, nLociAll + 1))
-    true_file = np.loadtxt(true_path, usecols=np.arange(1, nLociAll + 1))
-
-    metrics = dict(get_hap_switch_error_metrics(called_file, true_file, nInd, nLociAll))
-
-    print(f"Switch error rate: {metrics['switch_error_rate']}")
-    print(f"Phase error (intra) rate: {metrics['phase_error_rate']}")
-    print(f"Uncalled rate: {metrics['uncalled_rate']}")
-    print(
-        "Proportion of genotypes wrongly called as homozygote: "
-        f"{metrics['wrong_homozygote_rate']}"
-    )
-    print(
-        "Proportion of genotypes correctly called as heterozygote: "
-        f"{metrics['correct_heterozygote_rate']}"
-    )
-    print(f"Homozygote count in true genotype: {metrics['homozygote_count']}")
-    print(f"Heterozygote count in true genotype: {metrics['heterozygote_count']}")
-    print(f"Homo to hetero ratio: {metrics['homo_to_hetero_ratio']}")
-
-
-def main():
-    calc_switch_error_rate()
-
-
-if __name__ == "__main__":
-    main()
